+++ src/callbacks/preGetCollectible.ts
@@ -1,3 +1,4 @@
+import * as reverseJusticeFix from "../features/optional/bugfix/reverseJusticeFix";
 import { betterDevilAngelRoomsPreGetCollectible } from "../features/optional/major/betterDevilAngelRooms/callbacks/preGetCollectible";
 
 export function main(
@@ -10,6 +11,12 @@ export function main(
     `MC_PRE_GET_COLLECTIBLE - itemPoolType: ${itemPoolType}, decrease: ${decrease}, seed: ${seed}`,
   );
   */
+
+  // This has to be before the "betterDevilAngelRooms" feature
+  const returnValue = reverseJusticeFix.preGetCollectible();
+  if (returnValue !== undefined) {
+    return returnValue;
+  }
 
   return betterDevilAngelRoomsPreGetCollectible(itemPoolType, decrease, seed);
 }